providers/litellm/pyproject.toml
[project]
name = "lab-litellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.82.0",
    "python-dotenv>=1.2.2",
]
